requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.6
django-appconf==1.0.6
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-imagekit==5.0.0
django-js-asset==2.1.0
pilkit==3.0
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
